Move alterative plan handling to the separate file
